+++ CHANGES.rst
@@ -8,6 +8,7 @@ in progress
 general
 -------
 - Add HTTP-to-MQTT protocol forwarder component, see :ref:`forward-http-to-mqtt`
+- Add Terkin PHP, a HTTP API library for :ref:`daq-php`
 
 bugfixes
 --------